backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
Django==5.2.13
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
idna==3.12
mysqlclient==2.2.8
PyJWT==2.12.1
requests==2.33.1
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
django-cors-headers==4.3.1
django-jazzmin==2.6.0
